moving Dynamo cache client to proper dynamo namespace
